requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
distlib==0.4.0
distro==1.9.0
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
filelock==3.19.1
git-filter-repo==2.47.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
openai==2.8.1
pillow==11.3.0
platformdirs==4.4.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.34.0
